feat: Store the image of the Drink using Cloudinary.

+++ server/src/drinks/drinks.service.ts
@@ -2,30 +2,49 @@ import { Injectable } from '@nestjs/common';
 import { CreateDrinkDto } from './dto/create-drink.dto';
 import { UpdateDrinkDto } from './dto/update-drink.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { CloudinaryService } from 'src/cloudinary/cloudinary.service';
+import { ERROR_RESPONSE } from 'src/common/error.handle';
+import { ErrorCustom } from 'src/common/error.custom';
 
 @Injectable()
 export class DrinksService {
-  constructor(private prisma: PrismaService) { }
+  constructor(private prisma: PrismaService,
+    private cloudinaryService: CloudinaryService,
+  ) { }
 
-  async create(createDrinkDto: CreateDrinkDto) {
-    const { drink_name, price } = createDrinkDto;
-    const newDrinks = await this.prisma.drink.create({
-      data: {
-        drink_name,
-        price
-      }
-    })
-    for (let drinkDetails of createDrinkDto.drink_details) {
-      const ingredientWeightInKg = drinkDetails.ingredient_weight / 1000;
-      await this.prisma.drinksDetails.create({
+  async create(createDrinkDto: CreateDrinkDto, image: Express.Multer.File) {
+    try {
+      const { drink_name, price } = createDrinkDto;
+
+      const imageUploadResult = await this.cloudinaryService.uploadFile(image);
+      const imageUrl = imageUploadResult.secure_url;
+
+      const newDrinks = await this.prisma.drink.create({
         data: {
-          drink_id: newDrinks.drink_id,
-          ingredient_id: drinkDetails.ingredient_id,
-          ingredient_weight: ingredientWeightInKg
-        }
-      })
+          drink_name,
+          price: +price,
+          image_url: imageUrl
+        },
+      });
+
+      for (let drinkDetails of createDrinkDto.drink_details) {
+        const ingredientWeightInKg = drinkDetails.ingredient_weight / 1000;
+        const a = await this.prisma.drinksDetails.create({
+          data: {
+            drink_id: newDrinks.drink_id,
+            ingredient_id: +drinkDetails.ingredient_id,
+            ingredient_weight: +ingredientWeightInKg
+          }
+        });
+      }
+      return newDrinks;
+    } catch (error) {
+      if (image && image.path) {
+        await this.cloudinaryService.deleteImage(image.path);
+      }
+      console.error(`Error creating new drink:`, error);
+      throw new Error(`Error creating new drink: ${error.message}`);
     }
-    return newDrinks;
   }
 
   findAll() {
@@ -62,7 +81,28 @@ export class DrinksService {
     return updateDrink;
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const removeDrinkDetails = await this.prisma.drinksDetails.findMany({
+      where: {
+        drink_id: id,
+      }
+    });
+    if (!removeDrinkDetails) {
+      throw new ErrorCustom(ERROR_RESPONSE.DrinksIsNotExisted);
+    }
+    const deletePromises = removeDrinkDetails.map(async item =>
+      await this.prisma.drinksDetails.delete({
+        where: {
+          drink_id_ingredient_id: {
+            drink_id: item.drink_id,
+            ingredient_id: item.ingredient_id
+          }
+        }
+      })
+    );
+    await Promise.all(deletePromises);
     return this.prisma.drink.delete({ where: { drink_id: id } });
+
   }
+
 }
